% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{deltaCt}
\alias{deltaCt}
\title{Calculate delta Ct values from CT values using a variety of methods}
\usage{
deltaCt(ct, method = c("global", "endogenous", "genorm", "normfinder"),
  HKs = NULL, group = NULL, ...)
}
\arguments{
\item{ct}{character, The file name to be read in}

\item{method}{Logical, whether or not some message are output}

\item{HKs}{Logical, controlled by}

\item{group}{}

\item{...}{}
}
\value{
deltaCT values
}
\description{
Calculate delta Ct values using global, endogenous, genorm or normfinder
}
\examples{
# readEDS()

}
\seealso{
\code{\link{batchReadEDS}}
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
\keyword{CT}
\keyword{deltaCT}
\keyword{genorm}
\keyword{normfinder}
