% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{dim.miRoar}
\alias{dim.miRoar}
\title{dim miRoar object}
\usage{
\method{dim}{miRoar}(x)
}
\arguments{
\item{x}{object to be subsetted}
}
\value{
Dimensions
}
\description{
dim miRoar object
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
