% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{.extractEDSFile}
\alias{.extractEDSFile}
\title{Extract and parse the contents of an EDS file}
\usage{
.extractEDSFile(in_path, out_path, verbose = FALSE)
}
\arguments{
\item{in_path}{The input path to an EDS file.}

\item{out_path}{The output path to extract an EDS file to}

\item{verbose}{Boolean, whether or not messages are printed to console}
}
\value{
The contents of the extract EDS file in the form of a Plate R6 object.
}
\description{
Currently this function is not yet complete, I still would like to spend time
delving into the other files to see how this information can be used in the wells.
}
