% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{.processAnalysisTable}
\alias{.processAnalysisTable}
\title{Process the EDS Analysis Table}
\usage{
.processAnalysisTable(data, colheads)
}
\arguments{
\item{data}{A tab-separated string containing the information for a particular well}

\item{colheads}{A string of column headers to apply to the cycle threshold values.}
}
\value{
A list of named components and their values for a given Well.
}
\description{
The EDS analysis table contains various rows where there are typically 3-4 rows per well.
Each row will contain a different set of information that needs to be handled
}
