% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{.safe.as.numeric}
\alias{.safe.as.numeric}
\title{Safely convert a string to a numeric}
\usage{
.safe.as.numeric(x)
}
\arguments{
\item{x}{A character vector to be converted}
}
\value{
A numeric vector or a character vector if it cannot be converted losslessly
}
\description{
Will convert empty strings or strings that contain a single whitespace to a 0.
Typically used internally
}
