% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getColnames}
\alias{getColnames}
\title{colnames miRoar object
colnames miRoar object}
\usage{
getColnames(x)
}
\arguments{
\item{x}{object to be subsetted}
}
\value{
colnames miRoar object
}
\description{
colnames miRoar object
colnames miRoar object
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
