% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getRownames}
\alias{getRownames}
\title{rownames miRoar object}
\usage{
getRownames(x)
}
\arguments{
\item{x}{object to be subsetted}
}
\value{
rownames miRoar object
}
\description{
rownames miRoar object
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
