% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{ncol.miRoar}
\alias{ncol.miRoar}
\title{ncol miRoar object}
\usage{
ncol.miRoar(x)
}
\arguments{
\item{x}{object to be subsetted}
}
\value{
subsetted miRoar object
}
\description{
ncol miRoar object
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
