% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nrow.miRoar}
\alias{nrow.miRoar}
\title{nrow miRoar object}
\usage{
nrow.miRoar(x)
}
\arguments{
\item{x}{object to be subsetted}
}
\value{
nrow miRoar object
}
\description{
nrow miRoar object
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
