% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.miRoar}
\alias{print.miRoar}
\title{Print miRoar object}
\usage{
\method{print}{miRoar}(x)
}
\arguments{
\item{x}{object to be printed}
}
\value{
printed miRoar object
}
\description{
Print miRoar object
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
