% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{processEDSFiles}
\alias{processEDSFiles}
\title{Process a set of EDS files}
\usage{
processEDSFiles(input_path, directory = TRUE, verbose = TRUE)
}
\arguments{
\item{input_path}{A number.}

\item{verbose}{A number.}
}
\value{
A number.
}
\description{
Process a set of EDS files
}
\examples{
add(1, 1)
add(10, 1)
}
