% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{readEDS}
\alias{readEDS}
\title{Read single EDS file}
\usage{
readEDS(file, verbose = TRUE, suppress = FALSE)
}
\arguments{
\item{file}{character, The file name to be read in}

\item{verbose}{Logical, whether or not some message are output}

\item{suppress}{Logical, controlled by}
}
\value{
The contents of the EDS file in a list
}
\description{
Extracts and compiles a single EDS file
}
\examples{
# readEDS()

}
\seealso{
\code{\link{batchReadEDS}}
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
\keyword{eds}
