% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inout.R
\name{readEDSFiles}
\alias{readEDSFiles}
\title{readEDSfiles  read in EDS format data}
\usage{
readEDSFiles(files)
}
\arguments{
\item{files}{character vector of file paths}
}
\description{
readEDSfiles  read in EDS format data
}
