% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{removeBadSignals}
\alias{removeBadSignals}
\title{Remove Bad Signals}
\usage{
removeBadSignals(x, perc = 1)
}
\arguments{
\item{x}{a miRoar object}

\item{perc}{a}
}
\value{
x
}
\description{
Remove Bad Signals
}
\examples{
# readEDSfile()

}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
