% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{selectHKwrapper}
\alias{selectHKwrapper}
\title{A wrapper for the selectHKs function}
\usage{
selectHKwrapper(
  x,
  method = c("geNorm", "NormFinder"),
  minNrHKs = 2,
  group = NULL,
  trace = T,
  na.rm = T,
  log = T
)
}
\arguments{
\item{x}{ct Values}

\item{method}{character, which method to use, either genorm or normfinder}

\item{minNrHKs}{Numeric, number of miRs}

\item{group}{factor, group variable for normFinder}

\item{trace}{IF messages are to be produced}

\item{na.rm}{if na values should be removed}

\item{log}{if values are in log space}
}
\value{
the factor to be substract from CT values
}
\description{
A wrapper to calculate the geometric mean from HK genes selected by genorm or normfinder algorithms
}
\seealso{
\code{\link{deltaCT}}
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
\keyword{CT}
\keyword{deltaCT}
\keyword{genorm}
\keyword{normfinder}
