% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{setBadSignalsToNA.miRoar}
\alias{setBadSignalsToNA.miRoar}
\title{Set Bad Signals to NA}
\usage{
\method{setBadSignalsToNA}{miRoar}(x, maxCT = 40, minCT = 0, ampVal = 0, conf.val = 0.8)
}
\arguments{
\item{x}{a miRoar object}

\item{maxCT}{a}

\item{minCT}{b}

\item{ampVal}{c}

\item{conf.val}{d}
}
\value{
x
}
\description{
Set Bad Signals to NA
}
\examples{
# readEDSfile()

}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
