% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subset.miRoar}
\alias{subset.miRoar}
\title{Subset miRoar object}
\usage{
\method{subset}{miRoar}(x, i, j)
}
\arguments{
\item{x}{object to be subsetted}

\item{i}{row index}

\item{j}{col index}
}
\value{
subsetted miRoar object
}
\description{
Subset miRoar object
}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
