% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{updateHistory.miRoar}
\alias{updateHistory.miRoar}
\title{Update History of miRoar class object}
\usage{
\method{updateHistory}{miRoar}(x, timepoint0, message)
}
\arguments{
\item{x}{character, The file name to be read in}

\item{timepoint0}{The timepoint as generated by Sys.time()}

\item{message}{The message you wish to store}
}
\value{
x with the history updated with the specified method.
}
\description{
Update History of miRoar class object
}
\examples{

}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
