% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miR.R
\name{Well}
\alias{Well}
\title{Well class}
\description{
Describes a single Well and relevant properties
}
\examples{
\dontrun{
example <- Well$new(name= '')
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Well-new}{\code{Well$new()}}
\item \href{#method-Well-print}{\code{Well$print()}}
\item \href{#method-Well-clone}{\code{Well$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Well-new"></a>}}
\if{latex}{\out{\hypertarget{method-Well-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Well$new(values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Well-print"></a>}}
\if{latex}{\out{\hypertarget{method-Well-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Well$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Well-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Well-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Well$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
