% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{deltaCt.miRoar}
\alias{deltaCt.miRoar}
\title{Calculate delta Ct values from CT values using a variety of methods}
\usage{
\method{deltaCt}{miRoar}(ct, method = c("global", "endogenous", "geNorm",
  "NormFinder"), HKs = NULL, group = NULL, ...)
}
\arguments{
\item{ct}{character, The file name to be read in}

\item{method}{Logical, whether or not some message are output}

\item{HKs}{Logical, controlled by}

\item{group}{Factor describing the groups each sample belongs to}

\item{...}{Other parameters I have forgotten}

\item{which}{Which Ct matrix should be normalised ('Ct', 'Crt' or 'CtAvg')}
}
\value{
deltaCT values
}
\description{
Calculate delta Ct values using global, endogenous, genorm or normfinder
}
\examples{
# readEDS()

}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
