% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{updateHistory.miRoar}
\alias{updateHistory.miRoar}
\title{Update History of miRoar class object}
\usage{
\method{updateHistory}{miRoar}(x, timepoint0, message)
}
\arguments{
\item{x}{character, The file name to be read in}

\item{timepoint0}{Logical, whether or not some message are output}

\item{message}{Logical, controlled by}
}
\value{
x
}
\description{
Update History of miRoar class object
}
\examples{
# readEDS()

}
\author{
Tyler Gorrie-Stone \email{tgorri@essex.ac.uk}
}
